model: remove CreateUserRequest
